requirements.txt
torch>=2.2.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
peft>=0.10.0
bitsandbytes>=0.43.0
trl>=0.9.0
scikit-learn>=1.4.0
Pillow>=10.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.0.0
numpy>=1.24.0
tqdm
pydicom
medmnist
huggingface_hub
gradio>=5.0.0
